Consolidate the actor controlapi handlers into one file
